Move location of directory granularity model and actually read it from the database
